// src/api/model/fullMessage.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 */
import type { FullMessageTags } from './fullMessageTags';

export interface FullMessage {
  /**
   * @minimum -128
   * @maximum 127
   */
  type: number;
  text: string;
  displayName: string;
  timestamp: string;
  id: string;
  tags: FullMessageTags;
  username: string;
  channel: string;
  raw: string;
}